version final

+++ gestor/src/main/java/com/floresta/gestor/service/materialService.java
@@ -4,10 +4,13 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import com.floresta.gestor.dto.materialDTO;
 import com.floresta.gestor.model.material;
 import com.floresta.gestor.repository.materialRepository;
+
+import jakarta.validation.Valid;
 
 @Service
 public class materialService {
@@ -19,10 +22,12 @@ public class materialService {
         this.repository = materialRepository;
     }
 
-   
-    public material guardarMaterial(materialDTO request) {
+    
+
+    @Transactional
+    public material guardarMaterial(@Valid materialDTO request) {
     	
-    	material response = material.builder()
+    	material nuevoMaterial = material.builder()
     			.fecha(request.getFecha())
     			.material(request.getMaterial())
     			.cantidad(request.getCantidad())
@@ -31,22 +36,31 @@ public class materialService {
     			.precioTotal(request.getPrecioTotal())
     			.build();
     	
-        return repository.save(response);
+        return repository.save(nuevoMaterial);
     }
 
-
+ 
+    @Transactional(readOnly = true) 
     public List<material> obtenerTodos() {
         return repository.findAll();
     }
 
-
+    @Transactional(readOnly = true) 
     public Optional<material> obtenerPorId(Long id) {
         return repository.findById(id);
     }
 
+    @Transactional
+    public boolean eliminarPorId(Long id) {
+    	
+    	if (!repository.existsById(id)) return false; 
 
-    public void eliminarPorId(Long id) {
-    	repository.deleteById(id);
+		  try {
+		    repository.deleteById(id);
+		    return true;
+		  } catch (org.springframework.dao.DataIntegrityViolationException e) {
+		    return false; 
+		  }
     }
